refactor: 생성한 설문 조회 -> survey/inquire
