+++ src/pages/surveys/show/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import ShowSurvey from "../../../components/surveys/participate/ShowSurvey";
+import ShowSurvey from "../../../components/surveys/inquire/created/ShowSurvey";
 
 export default function Show() {
   const { sharingKey } = useParams();
