+++ src/pages/surveys/showCardSurvey/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ShowCardSurvey from "../../../components/surveys/participate/ShowCardSurvey";
+import ShowCardSurvey from "../../../components/surveys/inquire/created/card/ShowCardSurvey";
 
 export default function ShowCard() {
   return <ShowCardSurvey></ShowCardSurvey>;
